31/8 8:35

+++ controllers/license.js
@@ -1,10 +1,14 @@
-const config = require("../utils/config");
 const License = require("../models/license");
 const { updateModuleAfterRemoveLicense } = require("../controllers/module");
-
+const jwt = require("jsonwebtoken");
+const config = require("../utils/config");
 exports.createLicense = async (req, res, next) => {
   try {
-    const { name, image } = req.body;
+    const token = req.token;
+    const decodeToken = jwt.verify(token, config.SECRET);
+    if(!decodeToken.id || !decodeToken.role) return res.status(403).json({message: "Token missing or invalid"});
+    if(decodeToken.role !== "ADMIN") return res.status(403).json({message: "Role is not allowed"});
+    const { name, image, description } = req.body;
     const checkLicense = await License.findOne({ name });
     if (checkLicense) {
       return res.status(500).json({
@@ -15,6 +19,7 @@ exports.createLicense = async (req, res, next) => {
     const license = await new License({
       name,
       image,
+      description,
     });
     const licenseSaved = await license.save();
     res.status(201).json(licenseSaved);
@@ -32,8 +37,12 @@ exports.getListLicenses = async (req, res, next) => {
   }
 };
 
-exports.deleLicenseById = async (req, res, next) => {
+exports.deleteLicenseById = async (req, res, next) => {
   try {
+    const token = req.token;
+    const decodeToken = jwt.verify(token, config.SECRET);
+    if(!decodeToken.id || !decodeToken.role) return res.status(403).json({message: "Token missing or invalid"});
+    if(decodeToken.role !== "ADMIN") return res.status(403).json({message: "Role is not allowed"});
     const licenseId = req.params.id;
     const deletedLicense = await License.findByIdAndRemove({ _id: licenseId });
     if (!deletedLicense) {
@@ -41,9 +50,7 @@ exports.deleLicenseById = async (req, res, next) => {
         message: "License not found",
       });
     }
-    const updateModuleAfterRemoveLicense = await updateModuleAfterRemoveLicense(
-      licenseId
-    );
+    await updateModuleAfterRemoveLicense(licenseId);
     res.status(200).json(deletedLicense);
   } catch (error) {
     next(error);
@@ -52,6 +59,10 @@ exports.deleLicenseById = async (req, res, next) => {
 
 exports.updateLicenseData = async (req, res, next) => {
   try {
+    const token = req.token;
+    const decodeToken = jwt.verify(token, config.SECRET);
+    if(!decodeToken.id || !decodeToken.role) return res.status(403).json({message: "Token missing or invalid"});
+    if(decodeToken.role !== "ADMIN") return res.status(403).json({message: "Role is not allowed"});
     const data = req.body;
     const dataId = req.params.id;
     const updatedLisence = await License.findByIdAndUpdate(
@@ -59,6 +70,7 @@ exports.updateLicenseData = async (req, res, next) => {
       {
         name: data.name,
         image: data.image,
+        description: data.description,
       },
       { new: true }
     );
