+++ controllers/state.js
@@ -1,5 +1,6 @@
 const config = require("../utils/config");
 const State = require("../models/state");
+const jwt = require("jsonwebtoken");
 const { updateModuleAfterRemoveState } = require("../controllers/module");
 
 exports.getAllStates = async (req, res, next) => {
@@ -13,6 +14,10 @@ exports.getAllStates = async (req, res, next) => {
 
 exports.createState = async (req, res, next) => {
   try {
+    const token = req.token;
+    const decodeToken = jwt.verify(token, config.SECRET);
+    if(!decodeToken.id || !decodeToken.role) return res.status(403).json({message: "Token missing or invalid"});
+    if(decodeToken.role !== "ADMIN") return res.status(403).json({message: "Role is not allowed"});
     const { name } = req.body;
     const check = await State.findOne({ name: name });
     if (check) {
@@ -32,6 +37,10 @@ exports.createState = async (req, res, next) => {
 
 exports.deleteStateById = async (req, res, next) => {
   try {
+    const token = req.token;
+    const decodeToken = jwt.verify(token, config.SECRET);
+    if(!decodeToken.id || !decodeToken.role) return res.status(403).json({message: "Token missing or invalid"});
+    if(decodeToken.role !== "ADMIN") return res.status(403).json({message: "Role is not allowed"});
     const stateId = req.params.id;
     const deletedState = await State.findByIdAndRemove({ _id: stateId });
     if (!deletedState) {
@@ -39,7 +48,7 @@ exports.deleteStateById = async (req, res, next) => {
         message: "The state was not found",
       });
     }
-    const test = await updateModuleAfterRemoveState(stateId);
+    await updateModuleAfterRemoveState(stateId);
     res.status(200).json(deletedState);
   } catch (error) {
     next(error);
@@ -48,6 +57,10 @@ exports.deleteStateById = async (req, res, next) => {
 
 exports.updateStateData = async (req, res, next) => {
   try {
+    const token = req.token;
+    const decodeToken = jwt.verify(token, config.SECRET);
+    if(!decodeToken.id || !decodeToken.role) return res.status(403).json({message: "Token missing or invalid"});
+    if(decodeToken.role !== "ADMIN") return res.status(403).json({message: "Role is not allowed"});
     const data = req.body;
     const dataId = req.params.id;
     const updatedState = await State.findByIdAndUpdate(
